+++ DAL/ViewModels/VM_TheTaiReport.cs
@@ -21,6 +21,6 @@ namespace DAL.ViewModels
         public DateTime Luotve { get; set; }
         public List<tbTheTaiChiTiet> lstTheTaiChiTiet { get; set; } = new List<tbTheTaiChiTiet>();
         public int TienPhatSinh { get; set; }
-        public int SoKMPhatSinh { get; set; }
+        public decimal SoKMPhatSinh { get; set; }
     } 
 }
